# doc/api.yaml
openapi: 3.0.0
info:
  title: Home Library Service
  description: Home music library service
  version: 1.0.0
servers:
  - url: /
components:
  schemas:
...
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Innuendo'
        year:
          type: integer
          minimum: 1
          example: 1991
        artistId:
          type: string
          format: uuid
          nullable: true
      required:
        - name
        - year
        - artistId
...
paths:
...
  /album:
    get:
      tags:
        - Album
      summary: Gets albums list
      description: Gets all library albums list
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
    post:
      tags:
        - Album
      summary: Adds new album
      description: Adds new album information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                year:
                  type: integer
                  minimum: 1
                  example: 1991
                artistId:
                  type: string
                  format: uuid
                  nullable: true
              required:
                - name
                - year
                - artistId
      responses:
        201:
          description: Album is created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        400:
          description: Bad request (body does not contain required fields or the data type of the fields does not match the required or album name is empty string / year is equal 0)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
  /album/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Album
      summary: Gets single album by id
      description: Gets single album by id
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        400:
          description: Bad request(id (albumId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
    put:
      tags:
        - Album
      summary: Updates album information
      description: Updates library album information by UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                year:
                  type: integer
                artistId:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        200:
          description: The album has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        400:
          description: Bad request (id (albumId) is invalid (not uuid) or body is empty or the data type of the fields does not match the required or album name is empty string / duration is equal 0)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
    delete:
      tags:
        - Album
      summary: Deletes album
      description: Deletes album from library
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request (id (albumId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
...
  /favs/album/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags:
        - Favorites
      summary: Adds album to the favorites
      description: Adds album to the favorites
      responses:
        201:
          description: Added successfully
        400:
          description: Bad request (id (albumId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        422:
          description: Album with id doesn't exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
    delete:
      tags:
        - Favorites
      summary: Deletes album from favorites
      description: Deletes album from favorites
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request (id (albumId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
...